put bucket metrics tests in seperate file
